requirements.txt
Django==4.1.6
environs==9.5.0
Pillow==9.4.0
django-admin-sortable2==2.1.4
django-tinymce==3.5.0